ENH: Updating generators and snakemake parameters

+++ benchmarks/effect_size.py
@@ -19,22 +19,23 @@ quiet=True
 keep_logger=True
 
 # simulation parameters
+regenerate_simulations = True
 num_samples = 50
 num_features = 200
 #mu_num = 8            # mean of the numerator taxa
 mu_null = 0            # mean of the common taxa
 #mu_denom = 2          # mean of the denominator taxa
-max_diff = 3           # largest separation between the normals
-min_diff = 0.5         # smallest separation between the normals
+max_diff = 4           # largest separation between the normals
+min_diff = 0.05        # smallest separation between the normals
 sigma = 1              # variance of the random effects distribution
-pi1 = 0.25             # percentage of the species
-pi2 = 0.25             # percentage of the species
-low = -3               # lower value for spectrum
-high = 3               # higher value for the spectrum
-spread = 3             # variance of unimodal species distribution
+pi1 = 0.1              # percentage of the species
+pi2 = 0.3              # percentage of the species
+low = -4               # lower value for spectrum
+high = 4               # higher value for the spectrum
+spread = 2             # variance of unimodal species distribution
 feature_bias = 1       # species bias
-alpha = 6              # global sampling depth
-seed = None            # random seed
+alpha = 8              # global sampling depth
+seed = 0               # random seed
 
 # benchmark parameters
 top_N = int((pi1 + pi2) * num_features)     # top hits to evaluate
@@ -42,7 +43,7 @@ intervals = 3
 benchmark = 'effect_size'
 category = 'G'
 reps = 2
-tools = ['t_test', 'mann_whitney']
+tools = ['t_test', 'mann_whitney', 'random_forest', 'ancom', 'pseudo_balances']
 
 # generate config file
 data = {'category': category,
@@ -56,24 +57,25 @@ data = {'category': category,
 with open(config_file, 'w') as yfile:
     yaml.dump(data, yfile, default_flow_style=False)
 
-if not os.path.exists(output_dir):
-    os.mkdir(output_dir)
-else:
-    shutil.rmtree(output_dir)
-    os.mkdir(output_dir)
+if regenerate_simulations:
+    if not os.path.exists(output_dir):
+        os.mkdir(output_dir)
+    else:
+        shutil.rmtree(output_dir)
+        os.mkdir(output_dir)
 
-# generate simulations
-for i, ef in enumerate(np.linspace(max_diff, min_diff, intervals)):
-    for r in range(reps):
-        mu_num = mu_null + ef
-        mu_denom = mu_null - ef
-        res = block_table(num_samples, num_features,
-                          mu_num, mu_null, mu_denom, sigma, pi1, pi2,
-                          low, high, spread, feature_bias,
-                          alpha, seed=seed)
-        table, metadata, feature_metadata, beta, theta, gamma = res
-        metadata['effect_size'] = ef
-        deposit(table, metadata, feature_metadata, i, r, output_dir)
+    # generate simulations with respect to effect size
+    for i, ef in enumerate(np.linspace(max_diff, min_diff, intervals)):
+        for r in range(reps):
+            mu_num = mu_null + ef
+            mu_denom = mu_null - ef
+            res = block_table(num_samples, num_features,
+                              mu_num, mu_null, mu_denom, sigma, pi1, pi2,
+                              low, high, spread, feature_bias,
+                              alpha, seed=seed)
+            table, metadata, feature_metadata, beta, theta, gamma = res
+            metadata[benchmark] = ef
+            deposit(table, metadata, feature_metadata, i, r, output_dir)
 
 
 cmd = ' '.join([
